// src/app/core/models/setup/request/setup-etablissement-request.model.ts
export interface SetupEtablissementRequest {

  code: string;

  libelle: string;

  sigle: string;

  adresse: string;

  boitePostale: string;

  telephone: string;

  mobile: string;

  email: string;

  siteWeb: string;

  logoFileUuid: string;

  description: string;

  schoolType: string;

  anneeCreation: number;

}